Update generator package
Updated the generator to inlcude a function for taking inventory and calculating how many items were sold, and the total outstanding balance owed by the location. Next update should include a "remove" function that removes items *without* adding to the total owed of the location, as well as an option to accept payment from the location. Will also need to add "master" locations for warehousing and distribution