Update generator.go - sales data
Added sales tracking information to items, both within a location and overall. Also added the ability to take payments at a location, keeping track of how much is paid in addition to the running total owed. Next update, I'm hoping to move some of this information to json and possibly csv, and the update after that should see me starting on the web app code